# public/api/openapi.yaml
openapi: 3.0.0
info:
  title: Book Review API
  description: API for managing books, reviews, and user interactions
  version: 1.0.0
  contact:
    email: y.kabir1@edu.salford.ac.uk
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        profilePicture:
          type: string
          nullable: true
          example: "/uploads/profile_pictures/john-doe.jpg"
        roles:
          type: array
          items:
            type: string
          example: ["ROLE_USER"]
        banned:
          type: boolean
          example: false
        readingList:
          type: array
          items:
            $ref: '#/components/schemas/Book'
          example: []
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        googleBookId:
          type: string
          example: "zyTCAlFPjgYC"
        title:
          type: string
          example: "The Great Gatsby"
        authors:
          type: array
          items:
            type: string
          example: ["F. Scott Fitzgerald"]
        thumbnail:
          type: string
          nullable: true
          example: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
        description:
          type: string
          nullable: true
          example: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
        pageCount:
          type: integer
          nullable: true
          example: 180
        categories:
          type: array
          items:
            type: string
          example: ["Fiction", "Classics", "Literary"]
        averageRating:
          type: number
          format: float
          nullable: true
          example: 4.5
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        content:
          type: string
          example: "A masterpiece of American literature. The prose is beautiful and the story is timeless."
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        user:
          $ref: '#/components/schemas/User'
          example: {"id": 1, "name": "John Doe"}
        book:
          $ref: '#/components/schemas/Book'
          example: {"id": 1, "title": "The Great Gatsby"}
        likesCount:
          type: integer
          example: 12
        likedByCurrentUser:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example: {"email": ["This value is not a valid email address."]}
        code:
          type: integer
          example: 400
      required:
        - success
        - message
        - code
  responses:
    UnauthorizedError:
      description: Unauthorized - User is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            code: 401
            errors: {}
    ForbiddenError:
      description: Forbidden - User does not have required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Access denied. Admin role required."
            code: 403
            errors: {}
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Resource not found"
            code: 404
            errors: {}
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation failed"
            code: 400
            errors:
              email: ["This value is not a valid email address."]
              password: ["This value is too short. It should have 8 characters or more."]
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Resource already exists"
            code: 409
            errors: {}
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "An unexpected error occurred"
            code: 500
            errors: {}
paths:
...
  /reviews:
    post:
      summary: Create a review
      tags:
        - Reviews
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: string
                  example: "zyTCAlFPjgYC"
                content:
                  type: string
                  example: "A masterpiece of American literature. The prose is beautiful and the story is timeless."
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
            example:
              bookId: "zyTCAlFPjgYC"
              content: "A masterpiece of American literature. The prose is beautiful and the story is timeless."
              rating: 5
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review created successfully"
              example:
                message: "Review created successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/ServerError'

    get:
      summary: Get all reviews
      tags:
        - Reviews
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
          example: "Gatsby"
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, rating]
          description: Sort field
          example: "createdAt"
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction
          example: "desc"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
          description: Items per page
          example: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 25
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total_pages:
                        type: integer
                        example: 3
                      has_next:
                        type: boolean
                        example: true
                      has_prev:
                        type: boolean
                        example: false
              example:
                data: [
                  {
                    id: 1,
                    content: "A masterpiece of American literature. The prose is beautiful and the story is timeless.",
                    rating: 5,
                    createdAt: "2023-05-15T14:30:00Z",
                    updatedAt: "2023-05-15T14:30:00Z",
                    user: {"id": 1, "name": "John Doe"},
                    book: {"id": 1, "title": "The Great Gatsby"},
                    likesCount: 12,
                    likedByCurrentUser: true
                  },
                  {
                    id: 2,
                    content: "A classic novel that everyone should read at least once.",
                    rating: 4,
                    createdAt: "2023-05-14T10:15:00Z",
                    updatedAt: "2023-05-14T10:15:00Z",
                    user: {"id": 2, "name": "Jane Smith"},
                    book: {"id": 1, "title": "The Great Gatsby"},
                    likesCount: 8,
                    likedByCurrentUser: false
                  }
                ]
                pagination:
                  total: 25
                  page: 1
                  limit: 10
                  total_pages: 3
                  has_next: true
                  has_prev: false
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /reviews/{id}:
    put:
      summary: Update a review
      tags:
        - Reviews
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "An absolute masterpiece of American literature. The prose is beautiful and the story is timeless."
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
            example:
              content: "An absolute masterpiece of American literature. The prose is beautiful and the story is timeless."
              rating: 5
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review updated successfully"
              example:
                message: "Review updated successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Delete a review
      description: Delete a review. This operation can be performed by the review author, an admin, or a moderator.
      tags:
        - Reviews
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review deleted successfully"
              example:
                message: "Review deleted successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User is not the review author, admin, or moderator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: "Access denied. You must be the review author, an admin, or a moderator to delete this review."
                code: 403
                errors: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
...